+++ nix/tools/release/default.nix
@@ -53,11 +53,11 @@ let
         name = "postgrest-release";
         docs = "Patch postgrest.cabal, CHANGELOG.md, tag and push all in one go.";
         args = [ "ARG_POSITIONAL_SINGLE([version], [Version to release], [pre])" ];
-        inRootDir = true;
+        workingDir = "/";
       }
       ''
         trap "echo You need to be on the main branch or a release branch to proceed. Exiting ..." ERR
-        [[ "$(git rev-parse --abbrev-ref HEAD)" =~ ^main$|^rel- ]]
+        [[ "$(git rev-parse --abbrev-ref HEAD)" =~ ^main$|^v[0-9]+$ ]]
         trap "" ERR
 
         trap "echo You have uncommitted changes in postgrest.cabal. Exiting ..." ERR